+++ Sources/Pippin/Extensions/Foundation/Double/Double+Random.swift
@@ -7,11 +7,9 @@
 
 import Foundation
 
-extension Double {
-
+public extension Double {
     /// - Returns: a random floating point number between 0.0 and 1.0, inclusive.
-    public static var random: Double {
+    static var random: Double {
         return Double(arc4random()) / 0xFFFFFFFF
     }
-
 }