+++ Sources/Pippin/Extensions/Foundation/Int/Int+Random.swift
@@ -7,8 +7,7 @@
 
 import Foundation
 
-extension Int {
-
+public extension Int {
     /**
      - Parameters:
         - min: the lower bound on the random number to generate, must be ≥ 0, which is its default value.
@@ -16,9 +15,8 @@ extension Int {
      - Warning: supplying `min` < 0 or `max` ≤ `min` results in a fatal error.
      - Returns: random integer between `min` and `max`
      */
-    public static func random(min: Int = 0, max: Int) -> Int {
+    static func random(min: Int = 0, max: Int) -> Int {
         precondition(min >= 0 && min < max)
         return Int(arc4random_uniform(UInt32((max - min) + 1))) + min
     }
-
 }