+++ Sources/Pippin/Extensions/UIKit/UIColor/MoreColors.swift
@@ -8,34 +8,32 @@
 
 import UIKit
 
-extension UIColor {
-
-    public static var darkGreen: UIColor {
+public extension UIColor {
+    static var darkGreen: UIColor {
         return UIColor(red: 0, green: 0.5, blue: 0, alpha: 1)
     }
 
-    public static var darkBlue: UIColor {
+    static var darkBlue: UIColor {
         return UIColor(red: 0, green: 0, blue: 0.5, alpha: 1)
     }
 
-    public static var darkRed: UIColor {
+    static var darkRed: UIColor {
         return UIColor(red: 0.7, green: 0, blue: 0, alpha: 1)
     }
 
-    public static var lightLightGray: UIColor {
+    static var lightLightGray: UIColor {
         return UIColor(red: 0.9, green: 0.9, blue: 0.9, alpha: 1)
     }
 
-    public static var lightBlue: UIColor {
+    static var lightBlue: UIColor {
         return UIColor(red: 0.7, green: 0.7, blue: 0.9, alpha: 1)
     }
 
-    public static var lightGreen: UIColor {
+    static var lightGreen: UIColor {
         return UIColor(red: 0.7, green: 0.9, blue: 0.7, alpha: 1)
     }
 
-    public static var placeholder: UIColor {
+    static var placeholder: UIColor {
         return UIColor.color(red: 199, blue: 199, green: 205, alpha: 255)
     }
-
 }