+++ Sources/Pippin/Extensions/UIKit/UIViewController/UIViewController+ChildViewControllers.swift
@@ -8,9 +8,8 @@
 
 import UIKit
 
-extension UIViewController {
-
-    public func fillWithChildViewController(childViewController: UIViewController) {
+public extension UIViewController {
+    func fillWithChildViewController(childViewController: UIViewController) {
         addNewChildViewController(newChildViewController: childViewController)
         childViewController.view.fillSuperview()
     }
@@ -20,16 +19,15 @@ extension UIViewController {
     /// - Parameters:
     ///   - childViewController: The `UIViewController` that will be contained.
     ///   - containerView: An optional subview of the parent `UIViewController` to which the child `UIViewController`'s `view` will be added as a subview. If one is not provided, the parent's `view` is used as container.
-    public func addNewChildViewController(newChildViewController childViewController: UIViewController, containerView: UIView? = nil) {
+    func addNewChildViewController(newChildViewController childViewController: UIViewController, containerView: UIView? = nil) {
         let resolvedView: UIView = containerView ?? view
         resolvedView.addSubview(childViewController.view)
         addChild(childViewController)
     }
 
     /// Call on the instance of the child view controller to remove
-    public func removeAsChildViewController() {
+    func removeAsChildViewController() {
         view.removeFromSuperview()
         removeFromParent()
     }
-
 }